Database.go refractor done
