+++ backend/controllers/auth_controller_test.go
@@ -1,0 +1,87 @@
+package controllers // Replace with your actual package name
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	// "backend/services"
+	// "backend/utils"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+)
+
+// Mock for services.HandleRegister
+type MockHandleRegister struct {
+	mock.Mock
+}
+
+func (m *MockHandleRegister) MockHandleRegister(r *http.Request) (interface{}, error, int) {
+	args := m.Called(r)
+	return args.Get(0), args.Error(1), args.Int(2)
+}
+
+// Mock for utils.SendErrorResponse
+func MockSendErrorResponse(w http.ResponseWriter, message string, err error, status int) {
+	w.WriteHeader(status)
+	w.Write([]byte(message))
+}
+
+// Mock for utils.SendJSONResponse
+func MockSendJSONResponse(w http.ResponseWriter, response interface{}, status int) {
+	w.WriteHeader(status)
+	w.Write([]byte("Success"))
+}
+
+func TestRegister(t *testing.T) {
+	// Create a new instance of our MockHandleRegister
+	mockHandler := new(MockHandleRegister)
+
+	// Save the original functions and restore them after the test
+	// originalHandleRegister := services.HandleRegister
+	// originalSendErrorResponse := utils.SendErrorResponse
+	// originalSendJSONResponse := utils.SendJSONResponse
+	// defer func() {
+	// 	services.HandleRegister = originalHandleRegister
+	// 	utils.SendErrorResponse = originalSendErrorResponse
+	// 	utils.SendJSONResponse = originalSendJSONResponse
+	// }()
+
+	// // Replace the original functions with our mocks
+	// services.HandleRegister = mockHandler.HandleRegister
+	// utils.SendErrorResponse = MockSendErrorResponse
+	// utils.SendJSONResponse = MockSendJSONResponse
+
+	// Test case 1: Successful registration
+	t.Run("Successful Registration", func(t *testing.T) {
+		// Arrange
+		req := httptest.NewRequest(http.MethodPost, "/register", nil)
+		w := httptest.NewRecorder()
+		mockHandler.On("HandleRegister", req).Return("Success", nil, http.StatusOK)
+
+		// Act
+		Register(w, req)
+
+		// Assert
+		assert.Equal(t, http.StatusOK, w.Code)
+		assert.Equal(t, "Success", w.Body.String())
+		mockHandler.AssertExpectations(t)
+	})
+
+	// Test case 2: Failed registration
+	t.Run("Failed Registration", func(t *testing.T) {
+		// Arrange
+		req := httptest.NewRequest(http.MethodPost, "/register", nil)
+		w := httptest.NewRecorder()
+		mockHandler.On("HandleRegister", req).Return(nil, assert.AnError, http.StatusBadRequest)
+
+		// Act
+		Register(w, req)
+
+		// Assert
+		assert.Equal(t, http.StatusBadRequest, w.Code)
+		assert.Equal(t, "Registration Failed", w.Body.String())
+		mockHandler.AssertExpectations(t)
+	})
+}
